FEAT: Major System Upgrade - AI Integration, Security Hardening & Stability Fixes
🚀 Key Highlights:
- Integrated Google Gemini AI (v1 Stable Endpoint) for robust Patent Analysis.
- Implemented 'Smart Auto-Discovery' for AI Models (Auto-fallback: Flash -> Pro).
- Upgraded Security Architecture:
  - Enforced strict JWT Authentication for all AI endpoints.
  - Resolved CORS Policy issues and 403 Forbidden errors.
  - Secured API Key handling (sanitization & trimming).

🔧 Backend Enhancements:
- Refactored 'AIAnalysisService.java' to use Stable v1 API (Fixed 404/400 errors).
- Optimized Database Entities (Filing, IPAsset, AIQueryHistory) for better performance.
- Added Hibernate Context fetching logic for smarter AI responses.
- Cleaned up project structure and removed legacy logs.

📂 Configuration:
- Added template 'application.properties' & 'service-account.json' for easier deployment setup.

✨ Frontend Sync:
- Aligned React Frontend API calls with secure Backend endpoints.
- Fixed 404/403 errors in Analysis Page history fetching.

Build Status: Stable ✅

+++ backend/src/main/java/com/project/backend/repository/FilingTrackerRepository.java
@@ -9,9 +9,9 @@ import java.util.Optional;
 @Repository
 public interface FilingTrackerRepository extends JpaRepository<FilingTracker, Long> {
     
-    // ✅ FIX: Define the missing method to resolve Java(67108964)
-    Optional<FilingTracker> findByUserIdAndIpAssetId(Integer userId, Integer ipAssetId);
+    // 🔴 Change 1: Integer userId -> Long userId
+    Optional<FilingTracker> findByUserIdAndIpAssetId(Long userId, Integer ipAssetId);
 
-    // Fetches all tracked assets for the Filing Tracker page
-    List<FilingTracker> findByUserIdOrderByTrackedAtDesc(Integer userId);
+    // 🔴 Change 2: Integer userId -> Long userId
+    List<FilingTracker> findByUserIdOrderByTrackedAtDesc(Long userId);
 }
